make some logic
